Collapse TnsWindowData in TnsInfo to optimize CoderInfo layout
- Collapse redundant TnsWindowData struct into TnsInfo, removing intermediate struct nesting and alignment padding.
- Reorder `aCoeffs` float array first in TnsFilterData for 4-byte member alignment.
- Reduce executable `.text` binary size by 931 bytes compared to baseline.
- Retain scalar 32-bit types for CoderInfo members and BitCode to ensure maximum execution speed in hot loops.

+++ libfaac/channels.c
@@ -173,7 +173,7 @@ static int WriteICS(BitStream *bs, CoderInfo *coder, bool commonWindow)
     AccumBegin(&acc, bs);
     for (int i = 0; i < coder->datacnt; i++) {
         if (coder->s[i].len > 0) {
-            AccumPutBits(&acc, coder->s[i].data, coder->s[i].len);
+            AccumPutBits(&acc, (uint32_t)coder->s[i].data, coder->s[i].len);
             bits += coder->s[i].len;
         }
     }
